requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
django-currentuser==0.8.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.2.3
